+++ src/app/app.routes.ts
@@ -1,8 +1,22 @@
 import { Routes } from '@angular/router';
+import { ROUTES } from './shared/constants/routes';
 
 export const routes: Routes = [
   {
     path: '',
-    loadChildren: () => import('./core/components/navigation-tabs/tabs/tabs.routes').then((m) => m.routes),
+    redirectTo:  ROUTES.LOGIN,
+    pathMatch: 'full',
   },
+  {
+    path: 'auth',
+    loadChildren: () => import('./core/auth/auth.module').then((m) => m.AuthModule),
+  },
+  {
+    path: 'admin',
+    loadChildren: () => import('./admin/admin.module').then((m) => m.AdminModule),
+  },
+  {
+    path: 'client',
+    loadChildren: () => import('./client/client.module').then((m) => m.ClientModule),
+  }
 ];