+++ src/app/shared/constants/routes.ts
@@ -1,0 +1,9 @@
+export enum ROUTES {
+    LOGIN = 'auth/login',
+    REGISTER = 'auth/register',
+    HOME = 'client/menu/home',
+    DETAIL_PRODUCT = 'client/detail-product',
+    SHOPPING_CART = 'client/shopping-cart',
+    LIST_PRODUCTS = 'admin/list-products',
+    CREATE_OR_UPDATE_PRODUCT = 'admin/create-or-update-product',
+}