+++ CHANGELOG.md
@@ -1,4 +1,33 @@
 Changelog
+## [1.8.0](https://github.com/darksoul333666/Straapberry/compare/v1.7.0...v1.8.0) (2024-07-28)
+
+
+### ✨ Features
+
+* create and update products added, handlers, showin toasts and spinners ([5f482f6](https://github.com/darksoul333666/Straapberry/commit/5f482f611b38d45baf40c074f108109be40d8e66))
+* deleting product logic added ([58029ed](https://github.com/darksoul333666/Straapberry/commit/58029ed38bac710a9675567ed348936490fdd662))
+* list products added ([77ab822](https://github.com/darksoul333666/Straapberry/commit/77ab8228886cfbba86d8e626d1ac451060fb6d95))
+* ux for create and update product ([dc7f407](https://github.com/darksoul333666/Straapberry/commit/dc7f40755ac95a168a2efffcb120df469058f08f))
+* UX products list added ([f2c3176](https://github.com/darksoul333666/Straapberry/commit/f2c317657d7f8ea2c7b826f6f4b15509fb34ecbf))
+
+## [1.7.0](https://github.com/darksoul333666/Straapberry/compare/v1.6.0...v1.7.0) (2024-07-28)
+
+
+### ✨ Features
+
+* errors handled in register and login ([63e3971](https://github.com/darksoul333666/Straapberry/commit/63e3971bbd2b248de1d03038c5f49e2f3a787017))
+* forms elements stylized ([d071d30](https://github.com/darksoul333666/Straapberry/commit/d071d3043fe076c502448bad72677511dd8649f5))
+* logic for register and login ([ff33cfe](https://github.com/darksoul333666/Straapberry/commit/ff33cfeb3b79ede2d4e2526cb7d2da2ddef4f290))
+* navigation in login according to role user ([d8f1802](https://github.com/darksoul333666/Straapberry/commit/d8f180293aa9c0dd088d00938c2a59b0dfb46f49))
+* routing changed for handle auth navigation ([d859e67](https://github.com/darksoul333666/Straapberry/commit/d859e67c08477c588d26c5b7deab377d1198d2cc))
+
+## [1.6.0](https://github.com/darksoul333666/Straapberry/compare/v1.5.0...v1.6.0) (2024-07-27)
+
+
+### ✨ Features
+
+* adding storage and functions to handle auth, login, and register of users ([fa60086](https://github.com/darksoul333666/Straapberry/commit/fa600860be63d8860420893e50f6e3627331dd97))
+
 ## [1.5.0](https://github.com/darksoul333666/Straapberry/compare/v1.4.0...v1.5.0) (2024-07-27)